// my-unsplash/client/src/context/ImageContext/actions/types.ts
export enum ImageActionTypes {
  ADD_IMAGE_REQUEST = '[REQUEST] ADD IMAGE',
  ADD_IMAGE_COMPLETE = '[COMPLETE] ADD IMAGE',
  DELETE_IMAGE_REQUEST = '[REQUEST] DELETE IMAGE',
  DELETE_IMAGE_COMPLETE = '[COMPLETE] DELETE IMAGE',
  FETCH_IMAGE_REQUEST = '[REQUEST] FETCH IMAGE',
  FETCH_IMAGE_COMPLETE = '[COMPLETE] FETCH_IMAGE',
}